pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oh-my-flask"
version = "0.1.0"
description = "Recipe-based CLI for scaffolding and installing personal dev tools"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
omf = "omf.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["omf"]

[tool.hatch.build.targets.wheel.force-include]
"recipes" = "omf/_bundled/recipes"
"skeletons" = "omf/_bundled/skeletons"
"integrations" = "omf/_bundled/integrations"
